hw4/submission_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.9
azure-storage-blob
azure-identity==1.18.0
azure-servicebus==7.12.3
azure-cosmos==4.7.0
gunicorn
aiohttp
httpx
